requirements.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
Pillow==11.1.0
celery==5.4.0
django-celery-beat==2.9.0
redis==5.0.8
drf-spectacular==0.28.0
django-filter==24.3
openpyxl==3.1.5
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
